requirements.txt
matplotlib
pandas
Pillow
plotly
seaborn
streamlit
